Bug fixes to elfutils
frysk-imports/elfutils/libelf/ChangeLog:
2008-03-13  Tim Moore  <timoore@redhat.com>

	* elf_getdata_rawchunk.c (elf_getdata_rawchunk): Cast offset and
	size computation to uint64_t.

frysk-imports/elfutils/libdwfl/ChangeLog:
2008-03-13  Tim Moore  <timoore@redhat.com>

	* dwfl_module_addrsym.c (dwfl_module_addrsym): Start min_label
	from 0 to allow sizeless symbols below the address. Choose the
	best sizeless symbol.

+++ frysk-imports/elfutils/libelf/elf_getdata_rawchunk.c
@@ -79,7 +79,7 @@ elf_getdata_rawchunk (elf, offset, size, type)
     }
 
   if (unlikely (size > elf->maximum_size
-		|| (off64_t) (elf->maximum_size - size) < offset))
+		|| (uint64_t)(elf->maximum_size - size) < (uint64_t)offset))
     {
       /* Invalid request.  */
       __libelf_seterrno (ELF_E_INVALID_OP);
